backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.8.0
django-environ==0.12.0
django-oauth-toolkit==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
gunicorn==23.0.0
whitenoise==6.7.0
psycopg[binary]==3.2.4
